% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova_as_code.R
\name{ancova.as.code.plots}
\alias{ancova.as.code.plots}
\title{Code for ANCOVA Plot}
\usage{
ancova.as.code.plots(backup, dataname, dvs, between, covar, ext = "Rmd")
}
\description{
Code for ANCOVA Plot
}
