% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancovaAsText.R
\name{ancova.pwc.as.text}
\alias{ancova.pwc.as.text}
\title{Pairwise comparisons as text}
\usage{
ancova.pwc.as.text(pwcs, between, p.adjust.method = "bonferroni")
}
\description{
Pairwise comparisons as text
}
