% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{ancova.pwc}
\alias{ancova.pwc}
\title{Pairwise Comparisons of Estimated Marginal Means for ANCOVA Test}
\usage{
ancova.pwc(
  data,
  dvs,
  between,
  covar,
  p.adjust.method = "bonferroni",
  dv.var = NULL,
  as.table = F,
  only.sig = F
)
}
\arguments{
\item{data}{a data.frame or list containing the data}

\item{dvs}{numeric columns with the dependent variables to be used in the anova test}

\item{between}{independent (between) variable in which perform the anova test}

\item{covar}{column with the covariate information}

\item{p.adjust.method}{method to adjust p values for multiple comparisons.}

\item{dv.var}{column with the information to classify observations}

\item{as.table}{logical value indicating that the result should be returned after to apply `get.ancova.pwc.table` function}

\item{only.sig}{logical; if TRUE, only statistical significant results will be tabulated}
}
\value{
A data.frame containing the results for the pairwise comparisons
}
\description{
This function provides a wrapper for rstatix::emmeans_test for the dependent variables `dvs`
}
