% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{ancovaApp}
\alias{ancovaApp}
\title{Shiny Application to Run Factorial ANCOVA Tests}
\usage{
ancovaApp()
}
\description{
This function performs run a shiny application in which you can conduct Factorial ANCOVA tests
}
