% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.pwc}
\alias{anova.pwc}
\title{Pair-wise Anova Test}
\usage{
\method{anova}{pwc}(
  data,
  dvs,
  between = c(),
  within = c(),
  p.adjust.method = "bonferroni",
  dv.var = NULL,
  as.table = F,
  only.sig = F,
  skewness = c()
)
}
\description{
Pair-wise Anova Test
}
