% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{anovaApp}
\alias{anovaApp}
\title{Shiny Application to Run Factorial ANOVA Tests}
\usage{
anovaApp()
}
\description{
This function performs run a shiny application in which you can conduct Factorial ANOVA tests
}
