% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_statistics.R
\name{descriptive_statistics}
\alias{descriptive_statistics}
\title{Descriptive Statistics}
\usage{
descriptive_statistics(data, dvs, ivs = c(), type = "common", dv.var = NULL)
}
\arguments{
\item{data}{a data.frame in which we will perform the descriptive statistics}

\item{dvs}{numeric columns with the dependent variables}

\item{ivs}{columns with the independent variables}

\item{type}{a character string specifying type of summary statistics}

\item{dv.var}{column with the information to classify observations}
}
\value{
A data.frame containing the results for the descriptive statistics
}
\description{
Compute summary of descriptive statistics for numeric variables.
}
