% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialAnovaAsText.R
\name{factorial.anova.as.text}
\alias{factorial.anova.as.text}
\title{Factorial ANCOVA test as text}
\usage{
factorial.anova.as.text(aovs, data, between, effect.size = "ges")
}
\description{
Factorial ANCOVA test as text
}
