% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorialAnovaAsText.R
\name{factorial.anova.pwc.as.text}
\alias{factorial.anova.pwc.as.text}
\title{Pairwise comparisons as text}
\usage{
factorial.anova.pwc.as.text(pwcs, between, p.adjust.method = "bonferroni")
}
\description{
Pairwise comparisons as text
}
