% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{get.ancova.pwc.table}
\alias{get.ancova.pwc.table}
\title{ANCOVA pairwise comparisons into data.frame}
\usage{
get.ancova.pwc.table(pwcs, only.sig = F)
}
\arguments{
\item{pwcs}{a list of pairwise comparisons}
}
\value{
A data.frame containing the results for the anova test
}
\description{
This function performs a transformation of ANCOVA pairwise comparisons into a data.frame
}
