% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{get.ancova.table}
\alias{get.ancova.table}
\title{ANCOVA test results into data.frame}
\usage{
get.ancova.table(aovs)
}
\arguments{
\item{aovs}{a list of anova tests}
}
\value{
A data.frame containing the results for the anova test
}
\description{
This function performs a transformation of ANCOVA test results into a data.frame
}
