% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{get.wilcoxon.pwc.table}
\alias{get.wilcoxon.pwc.table}
\title{Tabulate Wilcoxon's Pairwise Comparisons}
\usage{
get.wilcoxon.pwc.table(pwcs, only.sig = F)
}
\arguments{
\item{pwcs}{a data.frame with the results of Wilcoxon's pairwise comparisons}

\item{only.sig}{logical; if TRUE, only statistical significant results will be tabulated}
}
\value{
A data.frame containing the pairwise comparisons
}
\description{
Creates table of pairwise comparisons based on results from Wilcoxon's test
}
