% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutliers.R
\name{getOutliers}
\alias{getOutliers}
\title{Identifying Outliers Using BoxPlot Methods}
\usage{
getOutliers(data, dvs, ivs = c())
}
\arguments{
\item{data}{a data.frame in which we will perform the identification of outliers}

\item{dvs}{numeric columns with the dependent variables}

\item{ivs}{columns to be used as independent variable in which perform the identification of outliers}
}
\value{
A data.frame containing the results rstatix::identify_outliers with column var in which is indicating each dependent variable
}
\description{
This function provides a wrapper for rstatic::identify_outliers for multiple dependent variables dvs
}
