% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova_plots.R
\name{ggBarPlotAoC}
\alias{ggBarPlotAoC}
\title{ANCOVA Bar Plot}
\usage{
ggBarPlotAoC(
  data,
  dv,
  iv,
  aov,
  pwc,
  covar = NULL,
  pre.post = NULL,
  bar.width = 0.75,
  color = c(),
  show.errorbar = T,
  theme = c(),
  font.size = list(text.x = 10, text.y = 10, title.x = 12, title.y = 12),
  step.increase = 0.2,
  subtitle = c(),
  ylim = NA
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the ANOVA test}

\item{dv}{a column name of dependent variable}

\item{iv}{a column name of independent variable}

\item{aov}{the ANOVA statistical results returned by rstatix::anova_test}

\item{pwc}{the statistical results returned by a pairwise comparisons between groups from rstatix}

\item{covar}{a column name of covariance variable}

\item{pre.post}{a column name of time variable (pre- and post-test phases)}

\item{bar.width}{the numeric value for the bar plots}

\item{color}{a vector containing the colors in the box-plot}

\item{show.errorbar}{a boolean that indicates if the plot shows error bar}

\item{font.size}{the list with the font label sizes}

\item{step.increase}{the numeric vector to be used to minimize the overlap}

\item{subtitle}{the subtitle in the plot, use number to indicate the row from ANCOVA table}

\item{ylim}{the number that indicates the axis-y limit}
}
\description{
This function create bar plots to report results from ANCOVA (AoV) test.
}
