% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova_plots.R
\name{ggPlotAoC}
\alias{ggPlotAoC}
\title{You can change the color as follows:
 lp + ggplot2::scale_color_manual(values=c("blue","pink"))}
\usage{
ggPlotAoC(
  data,
  x,
  y,
  color = c(),
  aov,
  pwc,
  linetype = color,
  by = c(),
  addParam = c(),
  font.label.size = 14,
  step.increase = 0.25,
  palette = "jco",
  p.label = "p.adj.signif"
)
}
\description{
You can change the color as follows:
 lp + ggplot2::scale_color_manual(values=c("blue","pink"))
}
