% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_plots.R
\name{ggPlotAoV}
\alias{ggPlotAoV}
\title{ANOVA Plot}
\usage{
ggPlotAoV(data, x, y, color = c(), aov, pwc, linetype = color,
  by = c(), addParam = c(), font.label.size = 10,
  step.increase = 0.005)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the ANOVA test}

\item{x}{a character string containing the name of variable used as x-axis}

\item{y}{a character string containing the name of variable used as y-axis}

\item{color}{a vector or character string containing the name of variable used as color in the box-plot}

\item{aov}{the ANOVA statistical results returned by rstatix::anova_test}

\item{pwc}{the statistical results returned by a pairwise comparisons between groups from rstatix}

\item{linetype}{the character string indicating the column in data for changing linetype}

\item{by}{the character vector containing the columns in the data to be used as variable to generate the grouping panels}

\item{addParam}{the character vector with elements to be included in the plot (e.g. "dotplot", "jitter", "boxplot", "point", "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_mad", "median_range"); see ?desc_statby for more details.}

\item{font.label.size}{the integer value with the font label size}

\item{step.increase}{the numeric vector to be used to minimize the overlap}
}
\value{
A ggplot object with the ANOVA plot
}
\description{
This function create box plots to report results from ANOVA (AoV).
}
