% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_plots.R
\name{ggPlotTTest}
\alias{ggPlotTTest}
\title{Independent T-Test Plot}
\usage{
ggPlotTTest(data, x, y, tt, addParam = c(), font.label.size = 14)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the t-test}

\item{x}{a character string containing the name of variable used as x-axis}

\item{y}{a character string containing the name of variable used as y-axis}

\item{tt}{the statistical results returned by a t_test from rstatix}

\item{addParam}{the character vector with elements to be included in the plot (e.g. "dotplot", "jitter", "boxplot", "point", "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_mad", "median_range"); see ?desc_statby for more details.}

\item{font.label.size}{the integer value with the font label size}
}
\value{
A ggplot object with the t-test plot
}
\description{
This function create ggpubr box-plot to report result from an independent t-test.
}
