% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric_plots.R
\name{ggPlotWilcoxon}
\alias{ggPlotWilcoxon}
\title{Non Parametric Wilcoxon Plot}
\usage{
ggPlotWilcoxon(data, x, y, wt, addParam = c(), font.label.size = 14)
}
\description{
Non Parametric Wilcoxon Plot
}
