% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity.R
\name{homogeneity.test}
\alias{homogeneity.test}
\title{Homogeneity test for ANCOVA and ANOVA test}
\usage{
homogeneity.test(
  data,
  dvs,
  between = c(),
  within = c(),
  covar = NULL,
  dv.var = NULL,
  skewness = c()
)
}
\arguments{
\item{data}{a data.frame containing the variables in which performing the homogenity test}

\item{dvs}{a character vector containing the dependent variables}

\item{between}{a character vector containing the independent variable used between-subject}

\item{within}{a character vector containing the independent variable used within-subject}

\item{covar}{column with the variable to be used as covariance}

\item{dv.var}{column with the information to classify observations based on dependent variables}

\item{skewness}{a list of transformation to achieve normality}
}
\value{
A data frame containing the homogeneity test (levene's test of variances and Anova slopes in ancova)
}
\description{
This function performs a homogeneity test of variance.
}
