% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneity_test.R
\name{homogeneity_test}
\alias{homogeneity_test}
\title{Homogeneity Test of Variance}
\usage{
homogeneity_test(data, dvs, between = c(), within = c(),
  dv.var = NULL)
}
\arguments{
\item{data}{a data.frame containing the variables in which performing the homogenity test}

\item{dvs}{a character vector containing the dependent variables}

\item{between}{a character vector containing the independent variable used between-subject}

\item{within}{a character vector containing the independent variable used within-subject}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A data frame containing the homogeneity test (levene's test)
}
\description{
This function performs a homogeneity test of variance.
}
