% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_as_code.R
\name{ind.ttest.as.code.plots}
\alias{ind.ttest.as.code.plots}
\title{Code for independent t-Test Plot}
\usage{
ind.ttest.as.code.plots(backup, dataname, dvs, iv, ext = "Rmd")
}
\description{
Code for independent t-Test Plot
}
