% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ind_ttest}
\alias{ind_ttest}
\title{Independent T-Test}
\usage{
ind_ttest(
  data,
  dvs,
  iv,
  alternative = "two.sided",
  var.equal = FALSE,
  hedges.correction = FALSE,
  dv.var = NULL,
  as.list = FALSE
)
}
\arguments{
\item{data}{a data.frame in which we will perform the t-test}

\item{dvs}{numeric columns with the dependent variables to be used in the t-test}

\item{iv}{independent (between) variable in which perform the t-test}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{hedges.correction}{logical value indicating whether apply the Hedges correction}

\item{dv.var}{column with the information to classify observations}

\item{as.list}{logical value indicating that the result should be returned as list indicating by separate the t_test and cohens_d effect size}
}
\value{
A data.frame containing the results for the independent t_test or a list with the dataframe in t.test and the t_test with their effect sizes
}
\description{
This function provides a wrapper for rstatic::t_test for multiple dependent variables dvs
including their effect sizes
}
