% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_as_code.R
\name{kruskal.as.code.plots}
\alias{kruskal.as.code.plots}
\title{Code for Kruskal-Wallis Plot}
\usage{
kruskal.as.code.plots(backup, dataname, dvs, between, ext = "Rmd")
}
\description{
Code for Kruskal-Wallis Plot
}
