% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal.R
\name{kruskal.test}
\alias{kruskal.test}
\title{Kruskal-Wallis Rank Sum Test}
\usage{
kruskal.test(data, dvs, between, dv.var = NULL, as.table = F)
}
\arguments{
\item{data}{a data.frame or list containing the data}

\item{dvs}{numeric columns with the dependent variables to be used in the kruskal test}

\item{between}{independent (between) variable in which perform the kruskal test}

\item{dv.var}{column with the information to classify observations}

\item{as.table}{logical value indicating that the result should be returned as table or as list}
}
\value{
A data.frame containing the results for the kruskal test and their effect sizes
}
\description{
This function provides a wrapper for rstatix::kruskal_test for the dependent variables `dvs`
including their effect sizes
}
