% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{kruskalApp}
\alias{kruskalApp}
\title{Shiny Application to Run Kruskal-Wallis Rank Sum Tests}
\usage{
kruskalApp()
}
\description{
This function performs run a shiny application in which you can conduct Kruskal-Wallis rank sum tests
}
