% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_code.R
\name{linearity.as.code}
\alias{linearity.as.code}
\title{Code for Linearity Assessment Plot}
\usage{
linearity.as.code(backup, dataname, dvname, covarname, ivsnames, ext = "Rmd")
}
\description{
Code for Linearity Assessment Plot
}
