% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_code.R
\name{linearity_code}
\alias{linearity_code}
\title{Code for Linearity Assessment Plot}
\usage{
linearity_code(backup, dataname, dvname, covarname, ivsnames, ext = "Rmd")
}
\description{
Code for Linearity Assessment Plot
}
