% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{normality.test.by.residual}
\alias{normality.test.by.residual}
\title{Normality Test of Residual Model}
\usage{
normality.test.by.residual(
  data,
  dvs,
  between = c(),
  within = c(),
  covar = NULL,
  wid = "row.pos",
  dv.var = NULL,
  skewness = c()
)
}
\arguments{
\item{data}{a data.frame or list of data containing the variables in which performing the normality test}

\item{dvs}{a character vector containing the dependent variables}

\item{between}{a character vector containing the independent variable used as between-subject}

\item{within}{a character vector containing the independent variable used as within-subject}

\item{covar}{a character indicating the column of covariate variable used in the normality test}

\item{dv.var}{column with the information to classify observations based on dependent variables}

\item{skewness}{a list of transformation to achieve normality}
}
\value{
A data frame containing the normality test
}
\description{
This function performs a normality test in a residual model.
}
