% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{normality_test}
\alias{normality_test}
\title{Normality Test}
\usage{
normality_test(x)
}
\arguments{
\item{x}{a numeric vector of data values}
}
\value{
A data frame containing the value of the normality statistic and its corresponding p.value
}
\description{
This function performs a normality test in a numeric vector.
}
