% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{normality_test_at}
\alias{normality_test_at}
\title{Normality Test}
\usage{
normality_test_at(dat, vars)
}
\arguments{
\item{dat}{a tibble data.frame containing the variables in which performing the normality test}

\item{vars}{a list of numeric columns in which performing the normality test}
}
\value{
A data frame containing the value of the normality statistic and its corresponding p.value
}
\description{
This function performs a normality test in diferents columns of vector.
}
