% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{normality_test_by_res}
\alias{normality_test_by_res}
\title{Normality Test of Residual Model}
\usage{
normality_test_by_res(
  data,
  dvs,
  between = c(),
  within = c(),
  covar = NULL,
  dv.var = NULL
)
}
\arguments{
\item{data}{a data.frame containing the variables in which performing the normality test}

\item{dvs}{a character vector containing the dependent variables}

\item{between}{a character vector containing the independent variable used as between-subject}

\item{within}{a character vector containing the independent variable used as within-subject}

\item{covar}{a character indicating the column of covariate variable used in the normality test}

\item{dv.var}{column with the information to classify observations based on dependent variables}

\item{ivs}{a character vector containing all the independent variables}
}
\value{
A data frame containing the normality test
}
\description{
This function performs a normality test in a residual model.
}
