% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova_plots.R
\name{oneWayAncovaBoxPlots}
\alias{oneWayAncovaBoxPlots}
\title{ANCOVA Box Plots}
\usage{
oneWayAncovaBoxPlots(
  data,
  dv,
  ivs,
  aov,
  pwcs,
  covar = c(),
  pre.post = c(),
  color = c(),
  theme = c(),
  step.increase = 0.05,
  subtitle = c(),
  ylim = NA
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the ANOVA test}

\item{dv}{a character string containing the name of variable used as x-axis}

\item{ivs}{a vector with character string containing the name of variable used as y-axis}

\item{aov}{the ANOVA statistical results returned by rstatix::anova_test}

\item{pwcs}{the statistical results returned by a pairwise comparisons between groups from rstatix}

\item{covar}{a column name of covariance variable}

\item{pre.post}{a column name of time variable (pre- and post-test phases)}

\item{color}{a vector or character string containing the name of variable used as color in the box-plot}

\item{theme}{the theme used in the box plot}

\item{step.increase}{the numeric vector to be used to minimize the overlap}

\item{subtitle}{the subtitle in the plot, use number to indicate the row from ANCOVA table}

\item{ylim}{the number that indicates the axis-y limit}
}
\description{
This function create box plots to report results from ANCOVA (AoV).
}
