% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova_plots.R
\name{oneWayAncovaPlots}
\alias{oneWayAncovaPlots}
\title{One-Way ANCOVA Plots}
\usage{
oneWayAncovaPlots(
  data,
  dv,
  ivs,
  aov,
  pwcs,
  addParam = c(),
  font.label.size = 14,
  step.increase = 0.25
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the one-Way ANOCVA test}

\item{dv}{a character string containing the variable used as y-axis}

\item{ivs}{a character vector containing the variables used as x-axis in one-Way ANCOVA test}

\item{aov}{the ANCOVA statistical results returned by rstatix::anova_test}

\item{pwcs}{the list of statistical results returned from pairwise comparisons between groups from rstatix}

\item{font.label.size}{the integer value with the font label size}

\item{step.increase}{the numeric vector to be used to minimize the overlap}
}
\value{
A list of ggplot objects with the Two-Way ANOVA plots
}
\description{
This functions returns a list of Plots related to one-Way ANCOVA test.
}
