% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{paired_ttest}
\alias{paired_ttest}
\title{Paired T-Test}
\usage{
paired_ttest(
  data,
  dvs,
  iv,
  alternative = "two.sided",
  var.equal = FALSE,
  hedges.correction = FALSE,
  dv.var = NULL,
  as.list = FALSE
)
}
\description{
Paired T-Test
}
