% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{remove_from_dataset}
\alias{remove_from_dataset}
\title{Remove Elements from DataSet}
\usage{
remove_from_dataset(data, to_remove, wid = "row.pos", dv.var = NULL)
}
\arguments{
\item{data}{a data.frame containing the dataset}

\item{to_remove}{a vector or list containing the elements to be removed from the dataset}

\item{wid}{a character vector containing the unique identifier for each row}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A data.frame with the eliminated dataset
}
\description{
Removes elements from a dataset based on a set of dependent variables
}
