% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_from_datatable}
\alias{remove_from_datatable}
\title{Remove Elements from a Data Table for Dependent Variables}
\usage{
remove_from_datatable(tbl, to_remove, wid = "row.pos", dv.var = NULL)
}
\arguments{
\item{tbl}{a data.frame containing the data table}

\item{to_remove}{a vector or list containing the elements to be removed from the data table}

\item{wid}{a character vector containing the unique identifier for each row}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A data.frame with the eliminated data table
}
\description{
Removes elements from a data table previously setting based on dependent variables
}
