% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_group_data}
\alias{remove_group_data}
\title{Remove elements for perforimg statiscits analysis}
\usage{
remove_group_data(pdat, dv, ivs, n.limit = 5)
}
\arguments{
\item{pdat}{a data.frame or data table with data to be removed}

\item{dv}{dependent variable}

\item{ivs}{independent variables}

\item{n.limit}{minimal value for a group}
}
\description{
Removes groups of elements from a data table in which we can not perform statistics analysis
}
