% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{round.pval}
\alias{round.pval}
\title{Rounded p-values to a number of digits}
\usage{
\method{round}{pval}(tbl, digits = 3, min = 0.001, cnames = c("p", "p.adj"))
}
\arguments{
\item{tbl}{a data.frame or data table with p-values to be rounded}

\item{digits}{number of digits for the p-values}

\item{min}{minimal value to be displayed as number}

\item{cnames}{a vector containing the columns names in which the p-values are displayed}
}
\value{
A data.frame with the p-values rounded
}
\description{
Removes elements from a data table previously setting based on dependent variables
}
