% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheirer.R
\name{scheirer.test}
\alias{scheirer.test}
\title{Scheirer Ray Hare test}
\usage{
scheirer.test(data, dvs, between, dv.var = NULL, as.table = F)
}
\arguments{
\item{data}{a data.frame or list containing the data}

\item{dvs}{numeric columns with the dependent variables to be used in the scheirer test}

\item{between}{independent (between) variable in which perform the scheirer test}

\item{dv.var}{column with the information to classify observations}

\item{as.table}{logical value indicating that the result should be returned as table or as list}
}
\value{
A data.frame containing the results for the scheirer test and their effect sizes
}
\description{
This function provides a wrapper for rcompanion::scheirerRayHare for the dependent variables `dvs`
including their effect sizes
}
