% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{scheirerApp}
\alias{scheirerApp}
\title{Shiny Application to Run Scheirer Ray Hare Tests}
\usage{
scheirerApp()
}
\description{
This function performs run a shiny application in which you can conduct Scheirer-Ray hare tests
}
