% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_utils.R
\name{set_dataset}
\alias{set_dataset}
\title{Set DataSet}
\usage{
set_dataset(data, dvs, wid = "row.pos", dv.var = "var")
}
\arguments{
\item{data}{a data.frame containing the initial data}

\item{dvs}{a character vector containing the dependent variables}

\item{wid}{a character vector containing the unique identifier for each row}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A data.frame with the created dataset
}
\description{
Creates a new dataset based on a set of dependent variables
}
