% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_datatable}
\alias{set_datatable}
\title{Settings Data Table for Dependent Variables}
\usage{
set_datatable(tbl, dvs, dv.var = "var")
}
\arguments{
\item{tbl}{a data.frame containing the initial data}

\item{dvs}{a character vector containing the dependent variables}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A data.frame with the created data table
}
\description{
Creates a new data table based on a set of dependent variables
}
