% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheirer_as_code.R
\name{srh.as.code.plots}
\alias{srh.as.code.plots}
\title{Code for Scheirer-Ray-Hare Plot}
\usage{
srh.as.code.plots(backup, dataname, dvs, between, ext = "Rmd")
}
\description{
Code for Scheirer-Ray-Hare Plot
}
