% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{symmetry.test}
\alias{symmetry.test}
\title{Symmetry test}
\usage{
symmetry.test(x)
}
\arguments{
\item{x}{a numeric vector of data values}
}
\value{
A data frame containing the value of the skewness and kurtosis values
}
\description{
This function performs a symmetry test in a numeric vector.
}
