% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{symmetry_test}
\alias{symmetry_test}
\title{Symmetry test}
\usage{
symmetry_test(x)
}
\arguments{
\item{x}{a numeric vector of data values}
}
\value{
A data frame containing the value of the skewness and kurtosis values
}
\description{
This function performs a symmetry test in a numeric vector.
}
