% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric_plots.R
\name{threeWayNonParamFactPlots}
\alias{threeWayNonParamFactPlots}
\title{Three-Way Factorial Non Parametric Plots}
\usage{
threeWayNonParamFactPlots(
  data,
  dv,
  ivs,
  non,
  pwcs,
  addParam = c(),
  font.label.size = 12,
  step.increase = 0.1,
  type = NULL
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the Three-Way Non Parametric test}

\item{dv}{a character string containing the variable used as y-axis}

\item{ivs}{a character vector containing the variables used as x-axis in Three-Way Non Parametric test}

\item{non}{the Non Parametric statistical results returned by rstatix::anova_test}

\item{pwcs}{the list of statistical results returned from pairwise comparisons between groups from rstatix}

\item{addParam}{the character vector with elements to be included in the plot (e.g. "dotplot", "jitter", "boxplot", "point", "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_mad", "median_range"); see ?desc_statby for more details.}

\item{font.label.size}{the integer value with the font label size}

\item{step.increase}{the numeric vector to be used to minimize the overlap}
}
\value{
A list of ggplot objects with the Three-Way Non Parametric plots
}
\description{
This functions returns a list of Box Plots related to Non Parametric test.
}
