% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_as_text.R
\name{ttest.as.text}
\alias{ttest.as.text}
\title{T-Test as textual report using APA style}
\usage{
ttest.as.text(
  t.test,
  ds,
  iv,
  var.equal = F,
  hedges.correction = T,
  lang = "en"
)
}
\description{
T-Test as textual report using APA style
}
