% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{ttestApp}
\alias{ttestApp}
\title{Shiny Application to Run Student's T-Tests}
\usage{
ttestApp()
}
\description{
This function performs run a shiny application in which you can conduct Student's t-tests
}
