% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_plots.R
\name{tTestPlots}
\alias{tTestPlots}
\title{Independent T-Test Plots}
\usage{
tTestPlots(
  data,
  dvs,
  iv,
  tts,
  addParam = c(),
  font.label.size = 14,
  dv.var = NULL
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the T-Test}

\item{dvs}{a character vector containing the dependent variables used as y-axis}

\item{iv}{a character vector containing the independent variable used as x-axis}

\item{tts}{the statistical results returned by rstatix::t_test}

\item{addParam}{the character vector with elements to be included in the plot (e.g. "dotplot", "jitter", "boxplot", "point", "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_mad", "median_range"); see ?desc_statby for more details.}

\item{font.label.size}{the integer value with the font label size}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A list of ggplot objects with the T-Test plots
}
\description{
This functions returns a list of Box Plots related to T-Test.
}
