% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_plots.R
\name{twoWayAnovaPlots}
\alias{twoWayAnovaPlots}
\title{Two-Way ANOVA Plots}
\usage{
twoWayAnovaPlots(
  data,
  dv,
  ivs,
  aov,
  pwcs,
  addParam = c(),
  font.label.size = 14,
  step.increase = 0.25
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the Two-Way ANOVA test}

\item{dv}{a character string containing the variable used as y-axis}

\item{ivs}{a character vector containing the variables used as x-axis in Two-Way ANOVA test}

\item{aov}{the ANOVA statistical results returned by rstatix::anova_test}

\item{pwcs}{the list of statistical results returned from pairwise comparisons between groups from rstatix}

\item{addParam}{the character vector with elements to be included in the plot (e.g. "dotplot", "jitter", "boxplot", "point", "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_mad", "median_range"); see ?desc_statby for more details.}

\item{font.label.size}{the integer value with the font label size}

\item{step.increase}{the numeric vector to be used to minimize the overlap}
}
\value{
A list of ggplot objects with the Two-Way ANOVA plots
}
\description{
This functions returns a list of Box Plots related to Two-Way ANOVA test.
}
