% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{twoWayNonParamFactPlots}
\alias{twoWayNonParamFactPlots}
\title{Two-Way Factorial Non-Parametric Plots}
\usage{
twoWayNonParamFactPlots(
  data,
  dv,
  ivs,
  non,
  pwcs,
  addParam = c(),
  font.label.size = 14,
  step.increase = 0.25,
  type = NULL,
  p.label = "p.adj.signif"
)
}
\arguments{
\item{data}{a data.frame containing the data in which performing the Two-Way  test}

\item{dv}{a character string containing the variable used as y-axis}

\item{ivs}{a character vector containing the variables used as x-axis in Two-Way  test}

\item{non}{the ANOVA statistical results returned by rstatix::anova_test}

\item{pwcs}{the list of statistical results returned from pairwise comparisons between groups from rstatix}

\item{addParam}{the character vector with elements to be included in the plot (e.g. "dotplot", "jitter", "boxplot", "point", "mean", "mean_se", "mean_sd", "mean_ci", "mean_range", "median", "median_iqr", "median_mad", "median_range"); see ?desc_statby for more details.}

\item{font.label.size}{the integer value with the font label size}

\item{step.increase}{the numeric vector to be used to minimize the overlap}

\item{p.label}{the label used for p-values}
}
\value{
A list of ggplot objects with the Two-Way  plots
}
\description{
This functions returns a list of Box Plots .
}
