% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_as_code.R
\name{wilcoxon.as.code.plots}
\alias{wilcoxon.as.code.plots}
\title{Code for Wilcoxon Signed-Rank Plot}
\usage{
wilcoxon.as.code.plots(backup, dataname, dvs, iv, ext = "Rmd")
}
\description{
Code for Wilcoxon Signed-Rank Plot
}
