% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxonText.R
\name{wilcoxon.pwc.as.text}
\alias{wilcoxon.pwc.as.text}
\title{Export as text a pairwise comparisons using Wilcoxon pairwise method}
\usage{
wilcoxon.pwc.as.text(
  pwcs,
  ds,
  between,
  p.adjust.method = "bonferroni",
  lang = "en"
)
}
\description{
Export as text a pairwise comparisons using Wilcoxon pairwise method
}
