% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcoxon.test}
\alias{wilcoxon.test}
\title{Wilcoxon Test}
\usage{
wilcoxon.test(
  data,
  dvs,
  iv,
  alternative = "two.sided",
  dv.var = NULL,
  as.list = FALSE
)
}
\arguments{
\item{data}{a data.frame or list containing the data}

\item{dvs}{numeric columns with the dependent variables to be used in the wilcoxon-test}

\item{iv}{independent (between) variable in which perform the wilcoxon-test}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{dv.var}{column with the information to classify observations}

\item{as.list}{logical value indicating that the result should be returned as list indicating by separate the wilcoxon.test and effect size}
}
\value{
A data.frame containing the results for the wilcoxon test and their effect sizes
}
\description{
This function provides a wrapper for rstatix::wilcox_test for the dependent variables `dvs`
including their effect sizes
}
