% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_code.R
\name{wilcoxon_plots_code}
\alias{wilcoxon_plots_code}
\title{Code for Wilcoxon Signed-Rank Plot}
\usage{
wilcoxon_plots_code(backup, dataname, dvs, iv, ext = "Rmd")
}
\description{
Code for Wilcoxon Signed-Rank Plot
}
