% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{wilcoxon_test}
\alias{wilcoxon_test}
\title{Wilcoxon Test}
\usage{
wilcoxon_test(
  data,
  dvs,
  iv,
  alternative = "two.sided",
  dv.var = NULL,
  as.list = FALSE
)
}
\arguments{
\item{data}{a data.frame in which we will perform the wilcoxon-test}

\item{dvs}{numeric columns with the dependent variables to be used in the wilcoxon-test}

\item{iv}{independent (between) variable in which perform the wilcoxon-test}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{dv.var}{column with the information to classify observations}

\item{as.list}{logical value indicating that the result should be returned as list indicating by separate the wilcoxon_test and effect size}
}
\value{
A data.frame containing the results for the independent wilcoxon_test or a list with the dataframe in wilcoxon.test and the wilcoxon_test with their effect sizes
}
\description{
This function provides a wrapper for rstatic::wilcox_test for multiple dependent variables dvs
including their effect sizes
}
