% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.R
\name{wilcoxonApp}
\alias{wilcoxonApp}
\title{Shiny Application to Run Wilcoxon Rank Sum Tests}
\usage{
wilcoxonApp()
}
\description{
This function performs run a shiny application in which you can conduct two-sample Wilcoxon tests
}
