% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winzorize.R
\name{winzorize}
\alias{winzorize}
\title{Winsorization}
\usage{
winzorize(
  dat,
  dv,
  ivs = NULL,
  covar = NULL,
  probs = c(0.05, 0.95),
  skewness = c()
)
}
\arguments{
\item{dat}{a data frame containing the values to be winzorized}

\item{dv}{column with the dependent variable}

\item{ivs}{columns with the independent variables}

\item{covar}{columns with the covariante}

\item{probs}{numeric vector of probabilities with values in [0,1] as used in quantile}

\item{skewness}{columns in which there were applied skewness transformation}
}
\value{
A data frame containing the column with dv values replaced
}
\description{
This function performs the replacement of extreme values by less extreme ones
in a dataframe for performing parametric tests
}
