% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aovText.R
\name{aov.pwc.as.text}
\alias{aov.pwc.as.text}
\title{Pairwise comparisons from ANCOVA or ANOVA as text}
\usage{
aov.pwc.as.text(pwcs, ds, between, p.adjust.method = "bonferroni", lang = "en")
}
\description{
Pairwise comparisons from ANCOVA or ANOVA as text
}
