% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinySettingOutliers.R
\name{getOutliers}
\alias{getOutliers}
\title{This function provides a wrapper for rstatic::identify_outliers for multiple dependent variables dvs}
\usage{
getOutliers(data, dvs, ivs = c(), covar = c(), is.extreme = T)
}
\description{
This function provides a wrapper for rstatic::identify_outliers for multiple dependent variables dvs
}
