% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_plots.R
\name{ggPlotPairedTTest}
\alias{ggPlotPairedTTest}
\title{Paired T-Test Plot}
\usage{
ggPlotPairedTTest(data, x, y, tt, id, font.label.size = 10)
}
\description{
This function create ggpubr box-plot to report result from a paired t-test.
}
