% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{normality_test_per_group}
\alias{normality_test_per_group}
\title{Normality Test per Groups}
\usage{
normality_test_per_group(data, dvs, ivs, dv.var = NULL)
}
\arguments{
\item{data}{a data.frame containing the variables in which performing the normality test}

\item{dvs}{a character vector containing the dependent variables}

\item{ivs}{character vector containing the independent variable used define the groups}

\item{dv.var}{column with the information to classify observations based on dependent variables}
}
\value{
A data frame containing the normality test
}
\description{
This function performs a normality test per groups.
}
