% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_text.R
\name{wilcoxon.pwc.as.text}
\alias{wilcoxon.pwc.as.text}
\title{Text for Pairwise Comparisons of Wilcoxon Signed-Rank Test}
\usage{
wilcoxon.pwc.as.text(
  pwcs,
  ds,
  between,
  p.adjust.method = "bonferroni",
  lang = "en"
)
}
\description{
Text for Pairwise Comparisons of Wilcoxon Signed-Rank Test
}
